Initial Commit
To use it you use python run_alicevison.py then the folder where all the generated data should go, then the folder with the images, then the folder with the alicevision binaries, and finally the number of images to scan in. So an example would be like "python run_alicevision.py build_files full Meshroom-2019.1.0/aliceVision/bin 41". Then the final model should be in the texturing folder within the generated data folder.